requirements.txt
altair
numpy
pandas
pydeck
streamlit
seaborn
matplotlib
psycopg2
etna-ts